update deploy script, separate question component, refactor
