+++ package.json
@@ -12,8 +12,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "eject": "react-scripts eject",
-    "predeploy": "npm build",
-    "deploy": "gh-pages -d build"
+    "deploy": "npm run build && gh-pages -d build"
   },
   "homepage": "http://hikurangi.github.io/maas-survey",
   "devDependencies": {
