+++ src/components/Survey.js
@@ -1,96 +1,49 @@
-import {
-  Box,
-  FormControl,
-  FormControlLabel,
-  // FormHelperText,
-  // FormLabel,
-  // List,
-  // ListItem,
-  Radio,
-  RadioGroup,
-}                           from '@material-u
-i/core'
-import React, { useState }  from 'react'
-
-import { scale, questions } from '../data'
-import Result               from './Result'
-import StatefulRadioGroup   from './StatefulRadioGroup'
+import { Box }             from '@material-ui/core'
+import React, { useState } from 'react'
+import { questions }       from '../data'
+import Question            from './Question'
+import Result              from './Result'
 
 const Survey = () => {
 
-  // const [state, setState] = useState()
+  const [state, setState] = useState({})
 
-  // function handleChange (e) {
-  //   console.log({targetName: e.target.name, targetValue: e.target.value})
-  //   setState({
-  //     ...state,
-  //     [e.target.name]: parseInt(e.target.value, 10)
-  //   })
-  // }
+  function handleChange (e) {
+    console.log({targetName: e.target.name, targetValue: e.target.value})
+    setState({
+      ...state,
+      [e.target.name]: parseInt(e.target.value, 10)
+    })
+  }
 
-  // const answers = Object.values(state)
-  // console.log({state, answers})
+  const answers = Object.values(state)
+  console.log({state, answers})
 
   return (
     <Box component='article'>
       <ol>
         {
           questions.map((item, index) => {
-            return <StatefulRadioGroup item={item} index={index} />
-            const radioGroupName = 'scale' + index
+            // index just in an array!
+            const name = 'scale' + index // TODO: unnecessary
+            const value = state[name] || ''
 
             return (
-              <li key={index}>
-                <p>{item}</p>
-                <FormControl component='fieldset'>
-                  <RadioGroup
-                    name={radioGroupName}
-                    onChange={handleChange}
-                    defaultValue={null}
-                    // value={state[radioGroupName]}
-                    row
-                  >
-                    {
-                      scale.map((item, index, array) => {
-              
-                        const displayIndex = index + 1
-                        const label        = displayIndex + '.'
-                        // const description  = getLabelDescription({ index, array })
-                        
-                        return (
-                          <FormControlLabel
-                            key={index}
-                            value={displayIndex}
-                            control={<Radio/>}
-                            label={label}
-                            labelPlacement={'start'}
-                          />
-                        )
-                      })
-                    }
-                  </RadioGroup>
-                </FormControl>
-              </li>
+              <Question
+                item={item}
+                index={index}
+                value={value}
+                name={name}
+                handleChange={handleChange}
+              />
             )
+
           })
         }
       </ol>
-      {/* <Result answers={answers} /> */}
+      <Result answers={answers} />
     </Box>
   )
 }
 
 export default Survey
-
-// function getLabelDescription ({ index, array }) {
-//   if (index === 0) {
-//     return 'Almost Always'
-
-//   } else if (index === array.length - 1) {
-//     return 'Almost Never'
-  
-//   } else {
-//     return ''
-  
-//   }
-// }
